repeat last reward after teminate state in make_target()
When Node is expanded after the termination state in MCTS, if reward and value are always 0, Node.value() in the search_path will approach 0.
By repeating the last reward r (while reversing the sign if it's a two-player game), Node.value() will approach r.
Without this, MuZero will miss the decisive move that decides the game (since it will be Q(s, a)->0).